+++ crates/stc_ts_file_analyzer/src/analyzer/generic/mod.rs
@@ -30,8 +30,7 @@ use tracing::{debug, error, info, span, trace, warn, Level};
 
 pub(crate) use self::{expander::ExtendsOpts, inference::InferTypeOpts};
 use crate::{
-    analyzer::{assign::AssignOpts, scope::ExpandOpts, Analyzer, Ctx, NormalizeTypeOpts},
-    ty::TypeExt,
+    analyzer::{scope::ExpandOpts, Analyzer, Ctx, NormalizeTypeOpts},
     util::{unwrap_ref_with_single_arg, RemoveTypes},
     VResult,
 };
@@ -678,100 +677,7 @@ impl Analyzer<'_, '_> {
 
                 debug!("({}): Inferred `{}` as {}", self.scope.depth(), name, dump_type_as_string(arg));
 
-                match inferred.type_params.entry(name.clone()) {
-                    Entry::Occupied(mut e) => {
-                        let _tracing = span!(
-                            Level::ERROR,
-                            "infer_type: type param",
-                            name = name.as_str(),
-                            new = tracing::field::display(&dump_type_as_string(arg)),
-                            prev = tracing::field::display(&dump_type_as_string(e.get()))
-                        )
-                        .entered();
-
-                        // Identical
-                        if e.get().type_eq(arg) {
-                            return Ok(());
-                        }
-
-                        if opts.append_type_as_union
-                            || self
-                                .assign_with_opts(
-                                    &mut Default::default(),
-                                    &arg.clone().generalize_lit(),
-                                    &e.get().clone().generalize_lit(),
-                                    AssignOpts {
-                                        span,
-                                        do_not_convert_enum_to_string_nor_number: true,
-                                        ignore_enum_variant_name: true,
-                                        ignore_tuple_length_difference: true,
-                                        ..Default::default()
-                                    },
-                                )
-                                .is_ok()
-                        {
-                            if (e.get().is_any() || e.get().is_unknown()) && !(arg.is_any() || arg.is_unknown()) {
-                                return Ok(());
-                            }
-
-                            if opts.ignore_builtin_object_interface && arg.is_builtin_interface("Object") {
-                                return Ok(());
-                            }
-
-                            debug!("Overriding");
-                            let new = if self
-                                .assign_with_opts(
-                                    &mut Default::default(),
-                                    arg,
-                                    e.get(),
-                                    AssignOpts {
-                                        span,
-                                        do_not_convert_enum_to_string_nor_number: true,
-                                        ..Default::default()
-                                    },
-                                )
-                                .is_ok()
-                            {
-                                arg.clone()
-                            } else {
-                                Type::new_union(span, vec![e.get().clone(), arg.clone()].freezed())
-                            };
-                            *e.get_mut() = new;
-                            return Ok(());
-                        }
-
-                        // If we inferred T as `number`, we don't need to add `1`.
-                        if self
-                            .assign_with_opts(
-                                &mut Default::default(),
-                                &e.get().clone().generalize_lit(),
-                                &arg.clone().generalize_lit(),
-                                AssignOpts {
-                                    span,
-                                    ..Default::default()
-                                },
-                            )
-                            .is_ok()
-                        {
-                            debug!("Ignoring the new type");
-
-                            return Ok(());
-                        }
-
-                        debug!("Cannot append");
-                        inferred.skip_generalization = true;
-
-                        if opts.use_error {
-                            inferred.errored.insert(name.clone());
-                        }
-                        return Ok(());
-                    }
-                    Entry::Vacant(e) => {
-                        let arg = arg.clone();
-
-                        e.insert(arg);
-                    }
-                }
+                self.upsert_inferred(span, inferred, name.clone(), arg, opts)?;
 
                 return Ok(());
             }
